fix: cast do native_conditioning para o dtype de compute na fase nativa
O tensor do TextEncodeQwenImageEdit chega fp32; sem o cast o wq do txtfusion
recebe float x bf16 (RuntimeError no primeiro step da fase nativa).

+++ nodes_reference_control.py
@@ -646,7 +646,7 @@ class CtxRushKrea2MultiRefControlledApply:
                 curve_power,
             )
             if multiplier <= 0.0 and native_context is not None:
-                ctx = native_context.to(device=x.device)
+                ctx = native_context.to(device=x.device, dtype=x.dtype)
                 if ctx.shape[0] != x.shape[0]:
                     ctx = ctx.repeat(x.shape[0], *([1] * (ctx.ndim - 1)))
                 refs = [
